[BE] refactor: Jwt 네이밍 변경 (#114)
Jwt 엔티티 -> Token
TokenInfo DTO -> Jwt
