+++ be/issue-tracker/src/main/java/org/presents/issuetracker/jwt/dto/Jwt.java
@@ -4,12 +4,12 @@ import lombok.Builder;
 import lombok.Getter;
 
 @Getter
-public class TokenInfo {
+public class Jwt {
 	private final String accessToken;
 	private final String refreshToken;
 
 	@Builder
-	private TokenInfo(String accessToken, String refreshToken) {
+	private Jwt(String accessToken, String refreshToken) {
 		this.accessToken = accessToken;
 		this.refreshToken = refreshToken;
 	}
